# user-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service provides two endpoints to manage Erply customers. It connects to a local PostgreSQL database and an external Erply service for user authentication and reading/writing customer data.",
        "title": "Erply service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/customer": {
            "post": {
                "description": "Add a new customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add a customer",
                "operationId": "add-customer",
                "parameters": [
                    {
                        "description": "Add Customer",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CustomerInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "502": {
                        "description": "Bad Gateway"
                    }
                }
            }
        },
        "/customer/{id}": {
            "get": {
                "description": "Get a customer by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a customer",
                "operationId": "get-customer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Customer"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.CustomerInput": {
            "type": "object",
            "required": [
                "email",
                "firstName",
                "lastName"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "twitterID": {
                    "type": "string"
                }
            }
        },
        "handler.Customer": {
            "description": "Details of a customer",
            "type": "object",
            "properties": {
                "EDI": {
                    "type": "string"
                },
                "GLN": {
                    "type": "string"
                },
                "PeppolID": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "address2": {
                    "type": "string"
                },
                "addressTypeID": {
                    "type": "integer"
                },
                "addressTypeName": {
                    "type": "string"
                },
                "birthday": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "colorStatus": {
                    "type": "string"
                },
                "companyName": {
                    "type": "string"
                },
                "companyTypeID": {
                    "type": "integer"
                },
                "country": {
                    "type": "string"
                },
                "countryID": {
                    "type": "string"
                },
                "credit": {
                    "type": "integer"
                },
                "creditCardLastNumbers": {
                    "type": "string"
                },
                "customerBalanceDisabled": {
                    "type": "integer"
                },
                "customerCardNumber": {
                    "type": "string"
                },
                "customerID": {
                    "type": "integer"
                },
                "customerType": {
                    "type": "string"
                },
                "deliveryTypeID": {
                    "type": "integer"
                },
                "doNotSell": {
                    "type": "integer"
                },
                "docuraEDIEnabled": {
                    "type": "integer"
                },
                "eInvoiceEmail": {
                    "type": "string"
                },
                "eInvoiceEnabled": {
                    "type": "integer"
                },
                "eInvoiceReference": {
                    "type": "string"
                },
                "ediType": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "emailEnabled": {
                    "type": "integer"
                },
                "emailOptOut": {
                    "type": "integer"
                },
                "euCustomerType": {
                    "type": "string"
                },
                "externalIDs": {
                    "type": "array",
                    "items": {}
                },
                "facebookName": {
                    "type": "string"
                },
                "factoringContractNumber": {
                    "type": "string"
                },
                "fax": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "flagStatus": {
                    "type": "integer"
                },
                "fullName": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "groupID": {
                    "type": "integer"
                },
                "groupName": {
                    "type": "string"
                },
                "homeStoreID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "integrationCode": {
                    "type": "string"
                },
                "isPOSDefaultCustomer": {
                    "type": "integer"
                },
                "lastModifierEmployeeID": {
                    "type": "integer"
                },
                "lastModifierUsername": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "mailEnabled": {
                    "type": "integer"
                },
                "mobile": {
                    "type": "string"
                },
                "operatorIdentifier": {
                    "type": "string"
                },
                "partialTaxExemption": {
                    "type": "integer"
                },
                "payerID": {
                    "type": "integer"
                },
                "paysViaFactoring": {
                    "type": "integer"
                },
                "personTitleID": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "posCouponsDisabled": {
                    "type": "integer"
                },
                "postalCode": {
                    "type": "string"
                },
                "referenceNumber": {
                    "type": "string"
                },
                "rewardPoints": {
                    "type": "integer"
                },
                "rewardPointsDisabled": {
                    "type": "integer"
                },
                "salesBlocked": {
                    "type": "integer"
                },
                "signUpStoreID": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "taxExempt": {
                    "type": "integer"
                },
                "twitterID": {
                    "type": "string"
                }
            }
        }
    }
}