# user-service/internal/domain/customer.go
package domain

// @model CustomerInput
type CustomerInput struct {
	FirstName string `json:"firstName" validate:"required"`
	LastName  string `json:"lastName" validate:"required"`
	Email     string `json:"email" validate:"required"`
	Phone     string `json:"phone"`
	TwitterID string `json:"twitterID"`
}